move context filter out of module
