+++ python_json_log_formatter/context_filter.pyi
@@ -1,0 +1,7 @@
+from logging import Filter, LogRecord
+from typing import Dict
+
+class ContextFilter(Filter):
+    def __init__(self, context: Dict[str, str]) -> None: ...
+    def update_context(self, context: Dict[str, str]) -> None: ...
+    def filter(self, record: LogRecord) -> bool: ...
